refactor: 불필요한 station 패키지 제거
