+++ src/main/java/com/swyp/mema/database/station/dto/response/subwayTime/SubwayTimeBasicRes.java
@@ -1,11 +1,10 @@
-package com.swyp.mema.domain.station.dto.response.subwayTime;
+package com.swyp.mema.database.station.dto.response.subwayTime;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+import lombok.Data;
 
 import java.util.ArrayList;
 import java.util.List;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-
-import lombok.Data;
 
 @Data
