+++ src/main/java/com/swyp/mema/database/station/dto/response/subwayTime/TotalSubwayTimeRes.java
@@ -1,9 +1,9 @@
-package com.swyp.mema.domain.station.dto.response.subwayTime;
-
-import java.util.List;
+package com.swyp.mema.database.station.dto.response.subwayTime;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
+
+import java.util.List;
 
 @Getter
 @AllArgsConstructor
