--- src/main/java/com/swyp/mema/domain/station/repository/TmpStationRepository.java
@@ -1,8 +1,0 @@
-package com.swyp.mema.domain.station.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import com.swyp.mema.domain.station.model.Station;
-
-public interface TmpStationRepository extends JpaRepository<Station, String> {
-}
