+++ src/main/java/com/swyp/mema/domain/voteLocation/dto/response/MidLocationRes.java
@@ -1,6 +1,6 @@
 package com.swyp.mema.domain.voteLocation.dto.response;
 
-import com.swyp.mema.domain.station.dto.response.subwayInfo.SingleStationRes;
+import com.swyp.mema.database.station.dto.response.subwayInfo.SingleStationRes;
 import lombok.Builder;
 import lombok.Getter;
